add past appointment saa success flow

+++ server/routes/appointments/appointmentsController.ts
@@ -55,7 +55,6 @@ import {
   DraftAppointment,
   DraftAppointmentBooking,
 } from '../serviceProviderReferrals/draftAppointment'
-import appointmentDecorator from '../../decorators/appointmentDecorator'
 import SupplierAssessment from '../../models/supplierAssessment'
 
 export default class AppointmentsController {
@@ -189,37 +188,37 @@ export default class AppointmentsController {
     if (appointmentScheduleDetails === null || !appointmentScheduleDetails.appointmentTime) {
       throw new Error('No scheduling details were filled in on the draft')
     }
-    // if (this.isAPastAppointment(appointmentScheduleDetails)) {
-    //   res.redirect(
-    //     `/service-provider/referrals/${req.params.id}/supplier-assessment/post-assessment-feedback/edit/${draft.id}/attendance`
-    //   )
-    // } else {
-    try {
-      await this.interventionsService.scheduleSupplierAssessmentAppointment(
-        res.locals.user.token.accessToken,
-        supplierAssessment.id,
-        draft.data
+    if (this.isAPastAppointment(appointmentScheduleDetails)) {
+      res.redirect(
+        `/service-provider/referrals/${req.params.id}/supplier-assessment/post-assessment-feedback/edit/${draft.id}/attendance`
       )
-    } catch (e) {
-      const interventionsServiceError = e as InterventionsServiceError
-      if (interventionsServiceError.status === 409) {
-        res.redirect(
-          `/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBookingId}/details?clash=true`
+    } else {
+      try {
+        await this.interventionsService.scheduleSupplierAssessmentAppointment(
+          res.locals.user.token.accessToken,
+          supplierAssessment.id,
+          draft.data
         )
-        return
+      } catch (e) {
+        const interventionsServiceError = e as InterventionsServiceError
+        if (interventionsServiceError.status === 409) {
+          res.redirect(
+            `/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBookingId}/details?clash=true`
+          )
+          return
+        }
+
+        throw e
       }
 
-      throw e
+      await this.draftsService.deleteDraft(draft.id, { userId: res.locals.user.userId })
+
+      const successfulRedirectPath = hasExistingScheduledAppointment
+        ? 'rescheduled-confirmation'
+        : 'scheduled-confirmation'
+
+      res.redirect(`/service-provider/referrals/${referralId}/supplier-assessment/${successfulRedirectPath}`)
     }
-
-    await this.draftsService.deleteDraft(draft.id, { userId: res.locals.user.userId })
-
-    const successfulRedirectPath = hasExistingScheduledAppointment
-      ? 'rescheduled-confirmation'
-      : 'scheduled-confirmation'
-
-    res.redirect(`/service-provider/referrals/${referralId}/supplier-assessment/${successfulRedirectPath}`)
-    // }
   }
 
   async showSupplierAssessmentAppointmentConfirmation(
@@ -465,16 +464,55 @@ export default class AppointmentsController {
         formError = data.error
         userInputData = req.body
       } else {
-        const updatedAppointment = await this.interventionsService.recordSupplierAssessmentAppointmentAttendance(
-          accessToken,
-          referralId,
-          data.paramsForUpdate
-        )
-        const redirectPath =
-          updatedAppointment.sessionFeedback?.attendance?.attended === 'no' ? 'check-your-answers' : 'behaviour'
-        res.redirect(
-          `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/${redirectPath}`
-        )
+        let basePath
+        let redirectPath
+        if (draftBookingId) {
+          const fetchResult = await this.fetchDraftAppointmentOrRenderMessage(req, res)
+          if (fetchResult.rendered) {
+            return
+          }
+          const { draft } = fetchResult
+          const draftAppointment = draft.data as DraftAppointment
+          if (draftAppointment) {
+            if (draftAppointment.sessionFeedback) {
+              draftAppointment.sessionFeedback.attendance.attended = data.paramsForUpdate.attended!
+              draftAppointment.sessionFeedback.attendance.additionalAttendanceInformation =
+                data.paramsForUpdate.additionalAttendanceInformation!
+            } else {
+              draftAppointment.sessionFeedback = {
+                attendance: {
+                  attended: data.paramsForUpdate.attended!,
+                  additionalAttendanceInformation: data.paramsForUpdate.additionalAttendanceInformation!,
+                },
+                behaviour: {
+                  behaviourDescription: null,
+                  notifyProbationPractitioner: null,
+                },
+                submitted: false,
+                submittedBy: null,
+              }
+            }
+          } else {
+            throw new Error('Draft appointment data is missing.')
+          }
+
+          await this.draftsService.updateDraft(draft.id, draft.data, { userId: res.locals.user.userId })
+
+          basePath = `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/edit/${draftBookingId}`
+          redirectPath =
+            draftAppointment.sessionFeedback.attendance.attended === 'no' ? 'check-your-answers' : 'behaviour'
+        } else {
+          const updatedAppointment = await this.interventionsService.recordSupplierAssessmentAppointmentAttendance(
+            accessToken,
+            referralId,
+            data.paramsForUpdate
+          )
+          basePath = `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback`
+          redirectPath =
+            updatedAppointment.sessionFeedback?.attendance?.attended === 'no' ? 'check-your-answers' : 'behaviour'
+        }
+
+        res.redirect(`${basePath}/${redirectPath}`)
       }
     }
 
@@ -497,12 +535,14 @@ export default class AppointmentsController {
     const { user } = res.locals
     const { accessToken } = user.token
     const referralId = req.params.id
+    const { draftBookingId } = req.params
 
     const [referral, supplierAssessment] = await Promise.all([
       this.interventionsService.getSentReferral(accessToken, referralId),
       this.interventionsService.getSupplierAssessment(accessToken, referralId),
     ])
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
+    const appointment = await this.getSupplierAssessmentAppointmentFromDraftOrService(req, res, supplierAssessment)
+
     if (appointment === null) {
       throw new Error('Attempting to add initial assessment behaviour feedback without a current appointment')
     }
@@ -515,14 +555,35 @@ export default class AppointmentsController {
         formError = data.error
         userInputData = req.body
       } else {
-        await this.interventionsService.recordSupplierAssessmentAppointmentBehaviour(
-          accessToken,
-          referralId,
-          data.paramsForUpdate
-        )
-        return res.redirect(
-          `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/check-your-answers`
-        )
+        let redirectUrl
+        if (draftBookingId) {
+          const fetchResult = await this.fetchDraftAppointmentOrRenderMessage(req, res)
+          if (fetchResult.rendered) {
+            return
+          }
+          const { draft } = fetchResult
+          const draftAppointment = draft.data as DraftAppointment
+          if (draftAppointment && draftAppointment.sessionFeedback) {
+            draftAppointment.sessionFeedback.behaviour.behaviourDescription = data.paramsForUpdate.behaviourDescription!
+            draftAppointment.sessionFeedback.behaviour.notifyProbationPractitioner =
+              data.paramsForUpdate.notifyProbationPractitioner!
+          } else {
+            throw new Error('Draft appointment data is missing.')
+          }
+
+          await this.draftsService.updateDraft(draft.id, draft.data, { userId: res.locals.user.userId })
+
+          redirectUrl = `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/edit/${draftBookingId}/check-your-answers`
+        } else {
+          await this.interventionsService.recordSupplierAssessmentAppointmentBehaviour(
+            accessToken,
+            referralId,
+            data.paramsForUpdate
+          )
+          redirectUrl = `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/check-your-answers`
+        }
+
+        res.redirect(redirectUrl)
       }
     }
     const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
@@ -534,19 +595,21 @@ export default class AppointmentsController {
       userInputData
     )
     const view = new BehaviourFeedbackView(presenter)
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
+    ControllerUtils.renderWithLayout(res, view, serviceUser)
   }
 
   async checkSupplierAssessmentFeedbackAnswers(req: Request, res: Response): Promise<void> {
     const { user } = res.locals
     const { accessToken } = user.token
     const referralId = req.params.id
+    const { draftBookingId } = req.params
 
     const [referral, supplierAssessment] = await Promise.all([
       this.interventionsService.getSentReferral(accessToken, referralId),
       this.interventionsService.getSupplierAssessment(accessToken, referralId),
     ])
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
+    const appointment = await this.getSupplierAssessmentAppointmentFromDraftOrService(req, res, supplierAssessment)
+
     if (appointment === null) {
       throw new Error('Attempting to check supplier assessment feedback answers without a current appointment')
     }
@@ -557,7 +620,8 @@ export default class AppointmentsController {
       appointment,
       serviceUser,
       referralId,
-      appointmentSummary
+      appointmentSummary,
+      draftBookingId
     )
     const view = new CheckFeedbackAnswersView(presenter)
 
@@ -568,17 +632,46 @@ export default class AppointmentsController {
     const { user } = res.locals
     const { accessToken } = user.token
     const referralId = req.params.id
+    const { draftBookingId } = req.params
+
     const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
+
+    const appointment = await this.getSupplierAssessmentAppointmentFromDraftOrService(req, res, supplierAssessment)
     if (appointment === null) {
       throw new Error('Attempting to submit supplier assessment feedback without a current appointment')
     }
+    if (draftBookingId) {
+      const fetchResult = await this.fetchDraftAppointmentOrRenderMessage(req, res)
+      if (fetchResult.rendered) {
+        return
+      }
+      const { draft } = fetchResult
+      const draftAppointment = draft.data as DraftAppointment
+      if (draftAppointment && draftAppointment.sessionFeedback) {
+        const data: CreateAppointmentSchedulingAndFeedback = {
+          appointmentTime: draftAppointment.appointmentTime,
+          durationInMinutes: draftAppointment.durationInMinutes,
+          appointmentDeliveryType: draftAppointment.appointmentDeliveryType,
+          sessionType: draftAppointment.sessionType,
+          appointmentDeliveryAddress: draftAppointment.appointmentDeliveryAddress,
+          npsOfficeCode: draftAppointment.npsOfficeCode,
+          appointmentAttendance: { ...draftAppointment.sessionFeedback.attendance },
+          appointmentBehaviour: { ...draftAppointment.sessionFeedback.behaviour },
+        }
+        await this.interventionsService.scheduleAndSubmitSupplierAssessmentAppointmentWithFeedback(
+          accessToken,
+          supplierAssessment.id,
+          data
+        )
+        await this.draftsService.deleteDraft(draft.id, { userId: res.locals.user.userId })
+      } else {
+        throw new Error('Draft appointment data is missing.')
+      }
+    } else {
+      await this.interventionsService.submitSupplierAssessmentAppointmentFeedback(accessToken, referralId)
+    }
 
-    await this.interventionsService.submitSupplierAssessmentAppointmentFeedback(accessToken, referralId)
-
-    return res.redirect(
-      `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/confirmation`
-    )
+    res.redirect(`/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/confirmation`)
   }
 
   async showSupplierAssessmentFeedbackConfirmation(req: Request, res: Response): Promise<void> {
@@ -1176,8 +1269,6 @@ export default class AppointmentsController {
     res: Response,
     supplierAssessment: SupplierAssessment
   ): Promise<InitialAssessmentAppointment | null> {
-    const { user } = res.locals
-    const { accessToken } = user.token
     const { draftBookingId, id } = req.params
     if (draftBookingId) {
       const fetchResult = await this.fetchDraftAppointmentOrRenderMessage(req, res)
